refactor: move User model to package model
